Fix bug with layout left()/up()
When moving down/right, we use the old cell size because we are moving
the start point outside of the existing cell. When we go up/right, we
are doing actually doing the inverse operation, moving backwards using
the new cell size. We still check to see if there's an old cell because
up()/down()/left()/right() should behave the same for the first cell.